I created the repository I am going to work in. I found an interesting article on visualising the madelbrot set using python. I think that would be a good place to start.
TO DO

Use the method in the medium article to generate an image of the mandelbrot set.

Think about what is neccessary for a zoom in on the mandelbrot set.

Think about how you will create this visualisation. Will you use moviepy or perhaps that other library that did not work very well before. I certainly would be interested in getting it working.