What have I done:
1. Edited polynomial_function.c so that it can read
the hex values in a file and break them into three
arrays specifically an array for each rgb value.

2. Added a Makefile

3. Added a file called hex_file which contains some hex colour values

TO DO

Fit each array to an even spacing of the domain [0,1]. Build an
interpolating spline for each of the arrays over that domain.

Find a way to pass the parameters of the spline out into a file.