Basically finished with this file. polynomial_fuction.c is able to output the necessary values into the files in "colour_parameters".
The files "hex_file" and "hex_file1" contain hex values to
make some okayish colour scales imao.

TO DO:

Start a new C file that creates images of the mandelbrot set and gradually zooms in.

- complex plane zoom
- escape time algorithm
- jpeg output.

Could make this whole colour choosing process more intuitive
and easy to do somehow. It is a pain to type in the hex values
at the moment. It is also annoying having to chuck a csv file
into the P5 online editor to examine the colour scale. To tell
the truth this whole process would probably be better if it was exported
to P5. That way I could have buttons and sliders to choose colours. But to be
honest a really polished colour gradient selector is a whole project in and
of itself. Could be cool though.