I have coded up a multiplier array and the center coordinates arrays for the zoom file. I have also coded up the base arrays that will be scaled and translated so to make the x and y values for the frames.
PROBLEM MAYBE?
I am worried that my growth multiplier values are not linear...
or more accurately are linear. I think the zoom may become
clunkier as the zoom approaches the final frame. I have a potential
solution to this involving a while loop and a growth factor to select x values.
The distance the center moves from the first center to the final should shrink
with each frame.