So I have created a file which give frame parameters for a off centre and centred zoom. I should be able to zoom in on any given point now to a high degree.
THINKING
Maybe I should try to save the pixel data, the particular x and y coordinates... I'm not too
sure. This is a real test of what C can do. But moving the zoom around to different points could be interesting.

I think I'm going to have to take a break from this for a bit until Wednesday.

TO DO
Make a single image of the mandelbrot set using your colouring algorithm.