Continued work on node js application.
Uint32 data type is now being used to communicate over the sockets. This
is ensuring the browser client is able to read socket input from the C
executable without issue.

I've added the client side code that the node server will ... serve. A
range of improvements can be made there. Client input to the node server
is a major goal.

Docker Networks could be used to allow the node server to communicate
with the container running the C executable.