Heavily refactored the webworker. It is now using a shared array buffer to efficiently store the incoming data. The client is now able to draw about as fast as the backend can supply it.
TO DO:
At the moment the program can only run one time before it resets. It will need some work to make it so multiple images can be explored in a single session.
Once the above is done it might be worth looking at how to move the docker container out to be a websocket connection to the node client. That way it will open up the possibility of having multiple backend machines.